python interface: add partial support for multiple inheritance
An isl_aff can be seen as a subclass of both an isl_multi_aff and
an isl_pw_aff.  In order to be able to export these types
to the Python bindings, some support for multiple inheritance
is therefore needed.

In particular, allow an isl type to have multiple superclasses.
If a call to a method of a superclass needs to be made,
then one of the superclasses is chosen fairly arbitrarily.
The assumption is that all the superclasses have some common
enclosing type such that in the worst case, the call will
be forwarded to that type.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>