python interface: add support for overloaded methods
Some functions in isl come in several variants depending
on the types of the arguments.  For example,
isl_pw_multi_aff_pullback_multi_aff and isl_pw_multi_aff_pullback_pw_multi_aff
perform essentially the same operation, but take a (final) argument
of a different type.  In a Python interface, there is no need to
provide several such methods.  Instead, the interface can check
the types of the arguments and can call the corresponding C function.

Add support for the construction of such overloaded methods.
In particular, if a function is marked __isl_overload, then it
is assumed that the final part of the function name corresponds
to the type of the final argument.  This suffix is removed
from the function name and a single Python method is created
for all functions with same resulting shortened name that
calls the appropriate function based on the type of the arguments
passed to the Python method.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>