python: make string parsing and returning Python 3 compatible
Python 3 requires explicit conversion from C strings (a.k.a. byte arrays)
to Python strings to ensure that the encoding choices are explicit.
Python 2 does not require these conversions, but also works without
problems if the Python 3 conversions are in the source code. Hence, this
commit changes string handling to use Python 3 style conversions.

For string conversion we need to choose an encoding type. As isl inputs
are all expected to be ASCII, we choose ascii as encoding.

Signed-off-by: Tobias Grosser <tobias@grosser.es>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>