cpp: explicitly manage isl_bool return values
Use isl::mangage(res) when returning isl_bool values, instead of relying
on implicit conversion.

Before this change the following code was generated:

  isl::boolean set::is_empty() const {
    auto res = isl_set_is_empty(get());
    return res;
  }

For this code, the return value of the function was implicitly
constructed from the value res of type isl_bool. As no implicit
constructor isl::boolean(isl_bool) was available, the constructor
isl::boolean(bool) was invoked and isl_bool was implicitly converted to
bool. As part of this conversion, isl_bool_error was converted to true,
which is incorrect. This issue is resolved by explicitly calling "return
isl::manage(res)".

Signed-off-by: Tobias Grosser <tobias@grosser.es>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>