cpp: add unit tests
This commit introduces unit tests to cover the isl C++ interface. In
comparison to the isl_test.c tests, no tables are used to achieve high
functional test coverage on all inputs. Instead, a test for each kind of
parameter, return value, constructor, or other different ways to use the
isl C++ interface is introduced to show that the C++ interface correctly
supports this functionality.

Strict C++11 conformance is used to ensure that the isl C++ interface
does not depend on any extensions.

Signed-off-by: Tobias Grosser <tobias@grosser.es>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>