interface: make language bindings depend on all relevant headers
The C++ language bindings were only made to depend on the all.h header
in 73b84e4a2a (cpp: generate C++ wrapper classes,
Sat Apr 8 05:12:50 2017 +0200) and this header itself only changes
very rarely.  In particular, the bindings were not made to depend
on the headers included by the all.h header.

Add a dependence to a dummy library that gets rebuilt whenever
one of the relevant headers is changed.
Rebuilding the C++ bindings whenever these headers change is important
for the unit tests that will be added in the next commit.

Also add such a dependence for the Python bindings.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>