interface: use vsnprintf instead of open_memstream for macosx support
The function open_memstream is used in combination with fprintf in order
to provide printf-like functionality on top of ostreams.
open_memstream, however, is not available on macosx.
The use of open_memstream and fprintf was therefore replaced by vsnprintf
and a manually allocated buffer (the size is determined via dry-running
vsnprintf).
This modification does not change the external behaviour but restores
macosx compatibility.

Signed-off-by: Alexander Matz <a.matz.1988@gmail.com>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>