interface: add support for callbacks not taking arguments
The current generator tacitly assumes that all callbacks
take their arguments.  Remove this assumption by explicitly
checking for __isl_take annotations.

Unfortunately, the memory management annotations of
arguments to functions pointers are not recorded by clang,
so they are not available for figuring out whether
a callback takes its arguments.
Look for a "__isl_take" string in the source instead.

As a side effect of this change, parentheses are now placed
around take arguments in the Python callback.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>