interface: add support for persistent callbacks
The callbacks in all currently exported functions
are only used during the call of the function that
have the callback as an argument.
There are however (currently not exported) functions,
such as isl_ast_build_set_at_each_domain, that only
store a callback in an object such that it can be used
by later operations on that object.
In the former case, the data needed for marshalling can
be stored on the stack, but in the latter case, this
data needs to survive for as long as the callback can
get called.  This means it needs to be stored
in the corresponding Python/C++ object.

The functions that set such a persistent callback
are detected as those that modify an object (return an object
of the same type as the first argument) and have exactly three
arguments, with the second being a callback (and the third
the corresponding user data).

In Python, only minor modifications are needed to support
persistent callbacks.  The marshalling data created
for callbacks in general simply needs to be referenced
from the object returned from the function that sets
the persistent callback.  Furthermore, any function that
modifies objects of this type needs to copy the reference
to the marshalling data since the underlying isl object
will have the same callback referencing that marshalling data.
Any method on objects of such types checks if any
of the persistent callbacks has failed, clears the state and
rethrows the exception.

In C++, it is not possible to store a reference/pointer
to marshalling data created on the stack.
Instead, the marshalling data for persistent callbacks
are turned into members of the class.
This data is not only copied in functions that modify
objects of that type, but also in the copy constructor and
the assignment operator.
Since there may be several copies of the data,
it is stored in a std::shared_ptr.
Since setting a persistent callback looks slightly
different from setting other callbacks, the C++ method
corresponding to a C function setting such a callback
gets a tailored implementation.
A call to release() on an object with any persistent
callbacks set is not allowed since it will be detached
from the marshalling data it is still referring to.
A call to copy() is still allowed, though.  However,
it is up to the user to ensure that the result
does not outlive the corresponding C++ object.

Reviewed-by: Oleksandr "Alex" Zinenko <ftynse@gmail.com>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>