interface/cpp.cc: extract out cpp_generator::get_return_type
The return type is currently only needed in the function header,
but when support for type based subclasses gets added,
it will also be needed in the function definition.
Moreover, the return type may then be different
from the type derived from the C function (and depend on "clazz".)
Extract a method for obtaining the return type
to make these changes easier to implement.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>