rename "noexceptions" C++ bindings to "checked"
The current bindings were placed in a "noexceptions" namespace
to allow for the introduction of another set of C++ bindings
that does use exceptions.
However, after a careful reexamination, it turns out that
the distinguishing difference between these two sets of bindings
is not so much that one uses exceptions and the other does not,
but more that one requires the user to add explicit error checking,
while the other does not.  In particular, the non-checking bindings
will throw an exception (if exceptions are available) or abort
(if they are not) if an error condition occurs, such that there
is no need for the user to check for error conditions explicitly
(other than catching exceptions if exceptions are available).

Note that the current bindings do not actually enforce
error checking, but this will be changed in an upcoming commit.
First the other set of bindings need to be introduced
such that users who cannot use exceptions but do not want
to add explicit error checking because they are OK
with the application aborting on error conditions
continue to have this option.
For now, simply rename the bindings to be able
to make the proper distinction when the other set
of bindings is introduced.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>