do not use inline namespace from checked C++ bindings
The next commit introduces conversion functions between
objects in the checked C++ bindings and objects
in the unchecked C++ bindings.
These conversion functions need to reference types
in both bindings at the same time, meaning that
the checked namespace cannot be inlined.

This means that a user who does not want to explicitly
mention the "checked" namespace (implying that only
the checked bindings are used), will have to do
the "inlining" herself.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>