checked C++ bindings: convert isl::checked::stat to class
Just like isl::boolean, isl::checked:stat should also keep track
of whether the user has checked whether any error has occurred.
This means that the object needs to keep track of whether
a check has occurred and that therefore isl::checked:stat needs
to be a class.

Use a temporary representation with constexpr static ok and error
fields such that no user code needs to be changed at this point.
This allows the change to the new interface to be performed
more gradually.
This representation does not allow keeping track of whether
errors have been checked and (especially) checking this state
at destruction time.
The ok and error fields will be removed after is_ok() and is_error()
methods have been introduced.

Note that using templates would allow the interface to be kept,
while also allowing testing whether the state has been checked.
However, since isl::checked::stat needs to be converted to a class
anyway, it might as well behave in a way that is similar
to isl::checked::boolean, which will be done in the next commit.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>