interface: handle functions setting an enum value
A function setting an enum value results in the generation
of several methods, one for setting each individual value.
The number of arguments of these methods is one less
compared to that of the function because the enum value set
is fixed for each method.

An alternative would be to pass in the original enum (in C++) or
to convert the enum into a C++/Python equivalent.
However, this would result in special handling of negative values
(perhaps not for setting an enum value, but definitely for later
support for querying the value), resulting in extra differences
between the two C++ bindings.