cpp_generator::print_method_header: extract out print_named_method_header
The extracted method will be used to print a method header using
a name that is different from its default name.
In particular, it will allow the header of a given method
to be printed using different names.  This will be used to print
two different headers for "get" methods,
one with the "get_" prefix and one without.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>