C++ bindings: automatically convert long arguments to isl::val
A user who wants to call a C++ method taking an isl::val
with some fixed integer value, first needs to construct
an isl::val from this integer value and the isl::ctx
of the object on which the method is called.
Since the isl::ctx is available in the object and
there is only one way of constructing an isl::val
from a fixed integer value, the C++ bindings
may as well perform the conversion automatically.

This is not needed for Python bindings, because any argument is
already automatically converted to the expected type and
because the context is implicit in the Python bindings.

For each combination of arguments that can be converted,
a separate overloaded method is introduced that performs
the conversion and then calls the method with the original
argument types.

A static list of names of functions that can be used
for such automatic conversion is maintained since
this is only needed for converting non-isl objects
to isl objects and there are only a few cases
where it makes sense to perform such a conversion
automatically.

The information about which isl types can be converted
from which non-isl types is derived from the list
of function names.  It might be possible to populate
this information directly, but it is unlikely
to be more convenient than the way it is extracted now.
The current approach also makes it easy to perform
some basic sanity checks on the conversion functions.

Reviewed-by: Manjunath Kudlur <manjunath@cerebras.net>
Signed-off-by: Sven Verdoolaege <sven@cerebras.net>